implementação stack route
