+++ src/app/(main)/_layout.tsx
@@ -1,0 +1,14 @@
+import { Stack } from 'expo-router';
+
+export default function RootLayout() {
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+        animation: 'fade',
+      }}
+    >
+      <Stack.Screen name="index" options={{ title: 'Início' }} />
+    </Stack>
+  )
+}
