+++ src/app/splash.tsx
@@ -1,0 +1,19 @@
+import { SplashScreen } from 'expo-router';
+import { useEffect, useState } from 'react';
+
+export function SplashScreenController() {
+  const [isLoading, setIsLoading] = useState(true);
+
+
+  useEffect(() => {
+    setTimeout(() => {
+      setIsLoading(false);
+    }, 3000);
+  }, []);
+
+  if (!isLoading) {
+    SplashScreen.hideAsync();
+  }
+
+  return null;
+}
